requirements.txt
Flask==3.0.3
gunicorn==20.1.0
Jinja2==3.1.4
pandas
pyjwt
Werkzeug
Authlib
google-cloud-storage
google-cloud-core
google-cloud-datastore
sdtp>=0.2.5
